record/client/requirements.txt
requests >= *
cassandra-driver >= *
couchdb >= *
couchbase >= *
elasticsearch >= *
happybase >= *
python-memcached >= *
pymongo >= *
mysql-connector-python >= *
neo4j-driver >= *
pyorient >= *
redis >= *

zmq >= *
numpy >= *
